+++ src/Config/SveaConfigurationProvider.php
@@ -1,12 +1,12 @@
 <?php
 require_once SVEA_REQUEST_DIR . '/Includes.php';
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
 
 /**
- * Description of ConfigurationProvider
+ * The class is used as Default to get the settings in SveaConfig
+ * The class can be used as an example when creating your own
+ * class implementing the ConfigurationProvider interface.
+ * The class should return the right authorization values
+ * and is called by the integration package.
  *
  * @author anne-hal
  */
@@ -17,7 +17,13 @@ class SveaConfigurationProvider implements ConfigurationProvider {
     public function __construct($enviromentConfig) {
         $this->conf = $enviromentConfig;
     }
-
+/**
+ * 
+ * @param type $type eg. INVOICE, PAYMENTPLAN, HOSTED
+ * @param type $country
+ * @return Username
+ * @throws Exception
+ */
     public function getUsername($type, $country) {
         $uType = strtoupper($type);
         $uCountry = strtoupper($country);
@@ -30,6 +36,13 @@ class SveaConfigurationProvider implements ConfigurationProvider {
         return $this->conf->conf['credentials'][$uCountry]['auth'][$uType]['username'];
     }
 
+    /**
+     * 
+     * @param type $type eg. INVOICE, PAYMENTPLAN, HOSTED
+     * @param type $country
+     * @return Password
+     * @throws Exception
+     */
     public function getPassword($type, $country) {
         $uType = strtoupper($type);
         $uCountry = strtoupper($country);
@@ -40,6 +53,13 @@ class SveaConfigurationProvider implements ConfigurationProvider {
         }
         return $this->conf->conf['credentials'][$uCountry]['auth'][$uType]['password'];
     }
+    /**
+     * 
+     * @param type $type eg. INVOICE, PAYMENTPLAN, HOSTED
+     * @param type $country
+     * @return ClientNumber
+     * @throws Exception
+     */
     public function getClientNumber($type, $country) {
         $uType = strtoupper($type);
         $uCountry = strtoupper($country);
@@ -50,6 +70,13 @@ class SveaConfigurationProvider implements ConfigurationProvider {
         }
         return $this->conf->conf['credentials'][$uCountry]['auth'][$uType]['clientNumber'];
     }
+    /**
+     * 
+     * @param type $type eg. INVOICE, PAYMENTPLAN, HOSTED
+     * @param type $country
+     * @return MerchantId
+     * @throws Exception
+     */
     public function getMerchantId($type, $country) {
         $uType = strtoupper($type);
         $uCountry = strtoupper($country);
@@ -60,7 +87,13 @@ class SveaConfigurationProvider implements ConfigurationProvider {
         }
         return $this->conf->conf['credentials'][$uCountry]['auth'][$uType]['merchantId'];
     }
-
+    /**
+     * 
+     * @param type $type eg. INVOICE, PAYMENTPLAN, HOSTED
+     * @param type $country
+     * @return Secret word
+     * @throws Exception
+     */
     public function getSecret($type, $country) {
         $uType = strtoupper($type);
         $uCountry = strtoupper($country);
@@ -71,7 +104,12 @@ class SveaConfigurationProvider implements ConfigurationProvider {
         }
         return $this->conf->conf['credentials'][$uCountry]['auth'][$uType]['secret'];
     }
-
+    /**
+     * 
+     * @param type $type eg. INVOICE, PAYMENTPLAN, HOSTED
+     * @return type
+     * @throws Exception
+     */
     public function getEndPoint($type) {
         $uType = strtoupper($type);
         if(array_key_exists($uType,$this->conf->conf['url']) == FALSE){